Axe the PublicRoomsManager and switch to passing the ModuleApi
